requirements.txt
# Python依赖
Flask==2.3.3
PyMySQL==1.1.0
pandas==2.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0